fixed small compile warning about size_t vs. int
git-svn-id: http://vtp.googlecode.com/svn/trunk@2621 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/vtString.h
@@ -255,7 +255,7 @@ protected:
 	void AllocBeforeWrite(int nLen);
 	void Release();
 	static void WIN_UNIX_STDCALL Release(vtStringData* pData);
-	static int WIN_UNIX_STDCALL SafeStrlen(pcchar lpsz);
+	static size_t WIN_UNIX_STDCALL SafeStrlen(pcchar lpsz);
 	static void FreeData(vtStringData* pData);
 	};
 
@@ -307,7 +307,7 @@ inline bool vtString::IsEmpty() const
 	{ return GetData()->nDataLength == 0; }
 inline vtString::operator pcchar() const
 	{ return m_pchData; }
-inline int WIN_UNIX_STDCALL vtString::SafeStrlen(pcchar lpsz)
+inline size_t WIN_UNIX_STDCALL vtString::SafeStrlen(pcchar lpsz)
 	{ return (lpsz == NULL) ? 0 : strlen(lpsz); }
 
 // vtString support (windows specific)
